Necessary changes for the new files layout
- The most important change is in the groupe structure :
From now on, the ens field is limited to the Type 0 universe.
This means that we limit ourselves to groups of first order.
I don't see a use case where we need to work with a higher order
group, but the reason for this change is that all the clumsy
coercion/conversion system that we made was completely broken
and could not deal with implicit coercions outside of the main
section where the groupe structure universe variable u was defined.
Now they work and we can work with the groupe "library" from the
outside :-)